% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputePrincipalAngles.R
\name{ComputePrincipalAngles}
\alias{ComputePrincipalAngles}
\title{Principal angles between subspaces}
\usage{
ComputePrincipalAngles(x, y)
}
\arguments{
\item{x, y}{Real matrices}
}
\description{
Function computes the principal angles between the subspaces
spanned by the columns of two matrices x and y
}
