% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbsp-dist.R
\name{sbsp.dist}
\alias{sbsp.dist}
\title{Distance between two subspaces}
\usage{
sbsp.dist(x, y, method = "grassmann", ...)
}
\arguments{
\item{x, y}{Real matrices}

\item{method}{The distance measure. See details.}
}
\description{
Function implements several distance measures between subspaces.
}
\details{
Allowable distance measures are
\itemize{
 \item{"grassmann": }{The geodesic distance on the Grassmann manifold.}
 \item{"fs": }{The Fubini-Study metric}
}
}
