% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbsp-expmap.R
\name{sbsp.expmap}
\alias{sbsp.expmap}
\title{Projection from the tangent space}
\usage{
sbsp.expmap(x, p = NULL)
}
\arguments{
\item{x}{A real matrix (a tangent vector)}

\item{p}{The point from whose tangent space to project x}
}
\value{
A real matrix.
}
\description{
Function projects a tangent vector onto the Grassmann manifold.
}
