% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbsp-interpolate.R
\name{sbsp.interpolate}
\alias{sbsp.interpolate}
\title{Interpolation between two subspaces}
\usage{
sbsp.interpolate(x, y, t)
}
\arguments{
\item{x, y}{Real matrices}

\item{t}{Interpolation parameter in \code{[0,infinity)}}
}
\description{
Function smoothly interpolates between two subspaces \code{x} and \code{y}.
The interpolation is parametrized by \code{t}, where \code{t=0} returns \code{x},
\code{t=1} returns \code{y}, and \code{t=.5} returns the mean of \code{x} and \code{y}
}
