% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbsp-logmap.R
\name{sbsp.logmap}
\alias{sbsp.logmap}
\title{Projection onto the tangent space}
\usage{
sbsp.logmap(x, p = NULL)
}
\arguments{
\item{x}{A real matrix}

\item{p}{The point on whose tangent space to project x}
}
\value{
A real matrix.
}
\description{
Function projects an element of the Grassmann manifold onto the tangent space at p
}
