% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbsp-mean.R
\name{sbsp.mean}
\alias{sbsp.mean}
\title{Compute the mean of a set of subspaces}
\usage{
sbsp.mean(x)
}
\arguments{
\item{x}{A list of matrices}
}
\value{
The mean of the matrices in \code{x}.
}
\description{
Function computes the mean of a set of subspaces
}
\details{
Function computes the mean of a set of symmetrix, positive-definite
matrices. Several methods are implemented:
}
